extract projects
